Separated whir code from utils
